separated functions in files. Dedicated files for joystick.{c,h}
