+++ software_package/in4073/pc_terminal/pc_terminal.c
@@ -7,26 +7,11 @@
  *------------------------------------------------------------
  */
 
-#include <ctype.h>
-#include <fcntl.h>
-#include <assert.h>
-#include <time.h>
-
-#include <stdio.h>
-#include <stdlib.h>
-#include <termios.h>
-#include <unistd.h>
-#include <string.h>
-#include <inttypes.h>
-#include <errno.h>
-#include <signal.h>
+#include "pc_terminal.h"
 
 
 #include "packet_constants.h"
 #include "joystick.h"
-
-#include <SDL.h>
-
 
 #define DEBUG
 //#define DEBUGCLK
@@ -40,57 +25,6 @@ unsigned short packet_id = 0;
 int 		fd;
 struct js_event js;
 
-
-JoystickData *JoystickData_create() {
-	JoystickData *jsdata = malloc(sizeof(JoystickData));
-	for (int i = 0; i < NBRAXES; i++)
-		(jsdata->axis)[i] = 0;
-	for (int i = 0; i < NBRBUTTONS; i++)
-		(jsdata->button)[i] = 0;
-	return jsdata;
-}
-
-void JoystickData_destroy(JoystickData *jsdata)
-{
-	free(jsdata);
-}
-
-void js_init()
-{
-	if ((fd = open(JS_DEV, O_RDONLY)) < 0) {
-		perror("jstest");
-		exit(1);
-	}
-	fcntl(fd, F_SETFL, O_NONBLOCK);
-}
-
-void js_getJoystickValue (JoystickData *jsdat)
-{
-		/* check up on JS
-		 */
-		while (read(fd, &js, sizeof(struct js_event)) == 
-		       			sizeof(struct js_event))  {
-
-			/* register data
-			 */
-			// fprintf(stderr,".");
-			switch(js.type & ~JS_EVENT_INIT) {
-				case JS_EVENT_BUTTON:
-					jsdat->button[js.number] = js.value;
-					break;
-				case JS_EVENT_AXIS:
-					jsdat->axis[js.number] = js.value;
-					break;
-			}
-		}
-		
-		if (errno != EAGAIN) {
-			perror("\njs: error reading (EAGAIN)");
-			exit (1);
-		}
-		
-
-}
 
 
 /*------------------------------------------------------------
@@ -321,7 +255,7 @@ int main(int argc, char **argv)
 
 	char isContinuing = 1;
 
-	js_init();
+	js_init(&fd);
 	JoystickData *jsdat = JoystickData_create();
 
 	struct timespec tp;
@@ -335,7 +269,7 @@ int main(int argc, char **argv)
 		// Response time is a bit variable. Latency can be percieved still.
 
 		// poll axes with raw-OS method
-		js_getJoystickValue(jsdat);
+		js_getJoystickValue(&fd, &js, jsdat);
 		for (int j = 0; j < NBRAXES; j++)
 		{
 			packet[AXISTHROTTLE + 2*j] = MSBYTE( jsdat->axis[j] );
