+++ software_package/in4073/pc_terminal/pc_terminal.h
@@ -1,0 +1,13 @@
+#include <ctype.h>
+#include <fcntl.h>
+#include <assert.h>
+#include <time.h>
+
+#include <stdio.h>
+#include <stdlib.h>
+#include <termios.h>
+#include <unistd.h>
+#include <string.h>
+#include <inttypes.h>
+#include <errno.h>
+#include <signal.h>
